Add OpNegate instruction

+++ src/rlox/value.rs
@@ -1,1 +1,1 @@
-pub type Value = u8;
+pub type Value = f32;
